requirements-dev.txt
wheel
flake8
pytest
black
twine
testcontainers_iris
requests<2.32.0